added authentication

+++ src/components/WorkoutDetails.js
@@ -1,4 +1,5 @@
 import { useWorkoutsContext } from '../hooks/useWorkoutsContext';
+import { useAuthContext } from '../hooks/useAuthContext';
 
 // date fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow';
@@ -6,12 +7,18 @@ import formatDistanceToNow from 'date-fns/formatDistanceToNow';
 const WorkoutDetails = ({ workout }) => {
 	const { dispatch } = useWorkoutsContext();
 
+	const { user } = useAuthContext();
+
 	const handleClick = async () => {
+		if (!user) {
+			return;
+		}
 		const response = await fetch(
 			'https://workout-buddy-0-1-0-backend.onrender.com/api/workouts/' +
 				workout._id,
 			{
 				method: 'DELETE',
+				headers: { Authorization: `Bearer ${user.token}` },
 			}
 		);
 		const data = await response.json();
